PDD-10-poc-create-headless-bi-api - Review comments fix
