+++ src/main/kotlin/uk/gov/justice/digital/hmpps/probationheadlessbipoc/config/WebClientConfiguration.kt
@@ -11,7 +11,6 @@ import java.time.Duration
 class WebClientConfiguration(
   @Value("\${hmpps-auth.url}") val hmppsAuthBaseUri: String,
   @Value("\${api.health-timeout:2s}") val healthTimeout: Duration,
-  @Value("\${api.timeout:20s}") val timeout: Duration,
 ) {
   // HMPPS Auth health ping is required if your service calls HMPPS Auth to get a token to call other services
   @Bean
