--- src/test/kotlin/uk/gov/justice/digital/hmpps/probationheadlessbipoc/integration/ExampleResourceTest.kt
@@ -1,59 +1,0 @@
-package uk.gov.justice.digital.hmpps.probationheadlessbipoc.integration
-
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-import java.time.LocalDate
-
-// TODO This test exists to support the HMPPS Typescript template and should be removed by the bootstrap process
-class ExampleResourceTest : IntegrationTestBase() {
-
-  @Nested
-  @DisplayName("GET /example/time")
-  inner class TimeEndpoint {
-
-    @Test
-    fun `should return forbidden if no token`() {
-      webTestClient.get()
-        .uri("/example/time")
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Test
-    fun `should return forbidden if no role`() {
-      webTestClient.get()
-        .uri("/example/time")
-        .headers(setAuthorisation())
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Test
-    fun `should return forbidden if wrong role`() {
-      webTestClient.get()
-        .uri("/example/time")
-        .headers(setAuthorisation(roles = listOf("ROLE_WRONG")))
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Test
-    fun `should return OK`() {
-      webTestClient.get()
-        .uri("/example/time")
-        .headers(setAuthorisation(roles = listOf("ROLE_TEMPLATE_KOTLIN__UI")))
-        .exchange()
-        .expectStatus()
-        .isOk
-        .expectBody()
-        .jsonPath("$").value<String> {
-          assertThat(it).startsWith("${LocalDate.now()}")
-        }
-    }
-  }
-}
